r move Messages container to dedicated module
